Rework gnu-efi path configuration
gnu-efi upstream as well as distros install the linker scripts into the
lib folder. Thus, there is no reason to configure those two directories
separately.

Remove GNUEFI_LDS_DIR and make sure that GNUEFI_LIB_DIR is set to the
right default, using based on "$CC -print-multi-os-directory". Fix the
reporting at the end of configure as well.

This finally fixes the detection of the gnu-efi location on SUSE.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>