env_api_fat: Fix problem with crc32 testing
If the crc was found to be wrong, the environment was used
normally despite, which caused problems in user tools.
Now, if an environment with an invalid crc is found, its working
copy is cleared. This way it can be automatically overwritten
by updates.

Also, the variable name 'oldenvs' is misleading, as it is an
array containing all current environments.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>