Bugfix: bgenv_get: Error codes overlap return values
If bgenv_get is called with a NULL pointer as data destination, it
returns the size of the buffer that must be provided by the user.

However, if the function fails, it also returns an error code greater
than zero, that overlaps possible values for the buffer size, so that
the user cannot distinguish failed and successful states.

To solve this, error codes are mapped to their negative counterparts,
i.e. bgenv_get returns -errno instead of errno.

To be symmetric to bgenv_set, the same is changed there.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>