bugfix: uservars: Fix deletion of deleted variables.
A user variable can be set to an empty string or it can be deleted. The
API distinguishes both cases using a USERVAR_TYPE_DELETED flag.

However, if a variable does not exist, and this flag is given, the
variable is wrongly created with the flag.

Fix the setter function so that if bgenv_find_uservar returns NULL, it
returns with no operation.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>