Make efibootguard compatible to gcc 8.1
GCC 8.1 has some additional `intelligent` tests, that do not allow the
length parameter in `strncat` or `strncpy` to be dependent on the
source. Despite the buffer being surely large enough, it complains with

error: 'strncpy' specified bound depends on the length of the
source argument [-Werror=stringop-overflow=].

Fact is, that all the functions that are used like this in the code
behave exactly as their counterpart `strcat` or `strcpy` respectively.
Hence, the solution is to just use strcat and strcpy.

Instead of malloc and strcpy, strdup can be used.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>