Fix memory leaks
Some gnu-efi functions which operate on DevicePaths internally
allocate memory, which must be freed afterwards.

AllocatePool and FreePool must never be called directly because
of the different calling convention. Always use the wrapper for
calling BS-> function pointers.

An empty line is added in the code to group code between
mmalloc and mfree.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>