Align init_array section to 4K
Signing tools, e.g. sbsigntool, stumble over the unalign init_array
section. According to [1], 4K was the default section alignment of
Visual Studio, and that made this the de-facto standard. Also gnuefi
aligns its sections this way.

As we do not want to fork our own linker script, use the assembly
fragment that places init_array_start at the beginning of that section
and align there.

[1] https://docs.microsoft.com/en-us/archive/msdn-magazine/2002/february/inside-windows-win32-portable-executable-file-format-in-detail

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>