Establish proper lifecycle management for bgenv API
When done with the with using it, bgenv should be finalized in order to
release resources allocated during bgenv_init and the later usage of
config_parts. Adjust the users accordingly.

Furthermore, bgenv_init should be made robust against multiple
succeeding invocations. They shall simply be ignored so that we do not
leak resources allocated on first init.

While at it, drop pointless zeroing of global var config_parts.

Reported-by: Tolga Hosgor <tolga.hosgor@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>