drivers/watchdog: Add support for Siemens SIMATIC IPC4x7E
While including iTCO support, like with other Intel chipsets, reset is
not wired on the IPC4x7E series. Therefore, we need to enable the
proprietary watchdog.

Detection requires consulting the SMBIOS table, looking for the
structure type 129 which contains OEM-specific information. In this
case, a so-called station ID tells us whether we have a regular Intel
100 series chipset or one on an IPC4x7E.

Another specialty is the SAFE_EN_N line that needs to be low in order to
allow the watchdog to trigger a hard reboot. This line is controlled by
GPP_A23. This is controlled via Community 0 registers that are located
via SBREG_BAR, which is under control of the S2SB bridge (0:1f.1). The
bridge device is hidden after boot, so we need to unhide it first in
order to read out SBREG_BAR. This still leaves us with a hard-coded
MMCFG base address, though, but this is safer than relying on some fixed
SBREG_BAR.

Based on original version by Johannes Pfefferl.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>