bg_setenv: Align -f parameter to bg_printenv
This switches bg_setenv -f to a file name, rather than just a directory.
The goal is to align bg_setenv and bg_printenv /wrt this argument.

In order to permit users to switch to the new scheme, detect the legacy
call, issue a warning, but still append the default environment file
name. This will eventually be dropped, though.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>