drivers/watchdog: add support for WDAT (ACPI) watchdogs
This driver parses ACPI tables and looks for a "WDAT entry. It
then programs the watchdog using the actions/instructions specified
by the BIOS (where register addresses and values are specified).
Unlike Linux, we not need to create (allocate) a dictionary of
actions as we may just parse the tables on the go to perform our
SET_COUNTDOWN and SET_RUNNING actions. Note that this driver is
tried before any others (the BIOS is expected to expose ACPI tables
with valid information). This driver was tested on SIMATIC IPC127E.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>
[Jan: address cppcheck warnings]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>