Avoid NULL dereference in FileDevicePathFromConfig
StrnCmp() dereferences both its CHAR16 pointer arguments.
If src is NULL which happens in the default case and if,
e.g., a CUSTOMLABEL is absent, StrnCmp() dereferences a
NULL pointer. Avoid this by calling StrnCmp() only if
src != NULL.

This is no change in behavior as devpath would not have
been set anyway.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>