ci: Migrate to github actions
Retire Travis which is no longer free, even for OSS, migrating to github
actions. This requires to open-code the Coverity integration which
turned out to be the simpler task.

The migration also brings CI integration for arm64. That was in fact the
hardest piece, both due to complications with Ubuntu (unhandy for cross-
building but only Linux image on github) and also glitches of EBG
itself.

Further change compared to travis: cppcheck is now pulled pre-built from
Ubuntu 21.04 (base image is still on 20.04, providing only an older
version). That removes the need to build and also cache that version
ourselves. For yet unknown reasons, the runtime of cppcheck is now twice
as long as under travis.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>