Build libebgenv.so shared library
Adapt the build system to build a shared library
libebgenv-${EBG_VERSION}.so symlinked to by a
libebgenv.so in addition to the static libebgenv.a.

This shared library may be used by external software
such as SWUpdate [1] wanting to modify EFI Boot Guard's
environment without having to statically link against
libebgenv.a, following the pattern of libubootenv [2].

Note that the bg_setenv utility deliberately remains
statically linked against libebgenv.a. for easy
inclusion in initramfs or rescue media and to not
break existing deployments. Alike, libebgenv.a is
still installed via make install (though distributions
may decide to not package it).
This may be deprecated and removed in future versions.

[1] https://github.com/sbabic/swupdate
[2] https://github.com/sbabic/libubootenv

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>