bg_setenv: simplify dump_envs
Previously, bg_printenv *always* implied verbosity (see line 670). The
verbosity flag was subsequently (mis)used to ensure the environments are
actually being dumped in the case of printenv. There is no need for such
a misuse though: if verbosity is false, then dump_envs is (mostly)
a no-op. In that case, do not even call the function or in other words,
move the if-clause where it belongs.

Finally, this allows us to leave verbosity as it is (i.e. unset) in the
case of bg_printenv. Now it is possible to give the verbose flag
a proper meaning in the case of printenv, i.e. the output could actually
be more verbose when the flag is set. This is not in the scope of this
commit, though.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>