tools: fix compile warnings
This fixes errors such as the following:

 147 | void dump_env(BG_ENVDATA *env, struct fields output_fields, bool raw)
     |      ^~~~~~~~ ../tools/bg_printenv.c: In function 'printenv_from_file': ../tools/bg_printenv.c:256:12: note: parameter passing for argument of type 'struct fields' changed in GCC 9.1

I would argue that the warning is harmless in this instance though.
Namely, the warning should only be relevant when building *shared objects*
which are called by code compiled with GCC < 9.1. However, bg_printenv and
bg_setenv are just binaries. Nevertheless, fixing it is trivial and
won't have any noticeable impact (I preferred to copy the struct in the
first place since it's cheaper than having the pointer indirection).

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>