Add stub for unified kernel images
This provides a simple UEFI stub which allows to form unified kernel
images, typically used for Linux. Besides the kernel itself, such images
can consist of an initrd, a static command line and device trees. All
artifacts can be combined into a single UEFI binary which may be signed
and will then be validated by the UEFI firmware in secure boot mode.

The device trees come into play when the firmware-provided one is not
compatible with the target kernel. The stub will match the device trees
of the unified image against the firmware version and select the one
that matches the compatible string.

Support for loading an initrd is limited to Linux kernels 5.8 or newer.

This stub is similar to what systemd provides but differs in being
simpler while providing the unique device tree matching functionality.

For CI, we need a few more cppcheck suppression.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>