bg_gen_unified_kernel: Make sections relocatable
If the stub image comes with a too small alignment, it's easy to grow
the PE header into the section data. This unfortunately happens on ARM
with current toolchains.

Resolve that by moving the sections up as needed when a new section is
added that causes an overflow. As the code has been prepared to account
for changing layouts between stub parsing, section adding and final
image writing, the actual changes are now small.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>