bg_gen_unified_kernel: Prepare for relocatable sections
We want to support moving the stub sections around in case the section
headers grow into them in the output file. To prepare for that, change
how the unified image is created: Use the section objects to obtain size
and specifically file position during image creation. That makes it
easier to move the data around while new sections are being added and
later - when writing out - refer to the right offsets.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>