bg_gen_unified_kernel: Adjust virtual address of text section
Current ARM toolchains use 8-byte alignment for the EFI binaries, and
that leads to the text section being loaded at a virtual address that
will immediate overlap with an expanded PE header like we produce when
adding sections. Resolve that by moving the text to address 0x1000 in
that case. This will give us sufficient room for a reasonable amount of
sections.

Changing the virtual address also requires to move AddressOfEntryPoint
and BaseOfCode in the PE header.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>