bg_gen_unified_kernel: Drop symbol table from image
Keeping it between the last stub section data and the first section we
append creates a problematic gap between section data in the file. This
is not explicitly allowed by the Authenticode specification for PE
files. That spec rather assumes that there is only extra data at the end
of the file which it demands to be hashed as well. The formula provided
in the spec to calculate the start and size of that extra data fails if
there are gaps between sections. While signing tools and EDK2 seem to be
fine with that, we are in a gray zone here with the generated image.

Avoid this by simply ripping out the symbol table before appending our
extra sections. We do that by tracking the final end of all section data
per PEHeaders, even across section additions in order to stay
consistent with this new (internal) API.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>