kernel-stub: Copy the kernel out of its section into a buffer
This is unavoidable as the UEFI firmware may decide to enforce the
section settings, making the region either executable (which it wasn't
so far) or writable, but not both at the same time. Seen with EDK2 on
arm64 so far, interestingly not on x86 yet. U-Boot didn't care at all.

Another issue of the current approach is that transplanting the kernel's
SectionAlignment into the unified image may also cause shifting of the
stub's sections if their original virtual address does not fulfill that,
and that shifting can break their relationship.

So this adopts the allocate and copy-out strategy of systemd-boot and
cleans up the generator script accordingly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>