Introduce validation of bgenv prior to its usage
The parsing of user variables assumes sane input so far and can be
mislead to out-of-bounds accesses, including writes. Address this by
always validating a bgenv after reading it from a partition or a file.
If an invalid bgenv is found, it is cleared to zero internally so that
the existing code will always operate against a sane state.

Include the CRC32 validation in the new helper as well which also
ensures that the checksum is tested when operating against a specific
file.

Reported by Code Intelligence.

Addresses CVE-2023-39950

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>