fix: correctly parse ustate in journal_process_action
This commit addresses an issue where the strtol function, which returns
a 'long', was assigned to an 'unsigned long' variable. As a result, the
error handling routine check 't == LONG_MIN' was ineffectual, leading to
an inability to correctly check for underflow.
Now the code uses the existing parsing logic 'parse_int' and verifies
that the parsed 'ustate' value falls within the correct range.

Note: the mismanagement of parse errors was not causing any issues
because 'ebg_env_setglobalstate' validates its input arguments
separately.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>