chore: ensure OPT macro fully initializes struct members
The usage of the OPT macro may result in uninitialized struct members
when applied to non-static variables. This commit mitigates this risk by
taking advantage of C99's designated initializers to ensure proper
initialization.
In the case of efibootguard, we're dealing with a static array of
structs, so each struct member is correctly initialized regardless of
this adjustment.

This fixes a clang compiler warning.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>