chore: introduce editorconfig for consistent coding styles
EditorConfig helps maintain uniform coding styles within a project,
regardless of the editor or IDE used. It uses a specific file format to
define coding styles, and text editor plugins to enforce those styles.
More details can be found at https://editorconfig.org.

With this commit, an EditorConfig file has been added to align with the
existing coding style in this repository. Most files adhere to a maximum
line length of 80 characters, except for configure.ac, where the line
length is closer to 100. This EditorConfig file will help ensure that
these styles are consistently followed in the future.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>