fix: correctly calculate FAT bit size based on cluster count
This commit addresses an issue in determining the bit size for FAT
filesystems. The previous approach was as follows:

1. Check the partition GUID against a hard-coded list of allowed values.
2. Assign 12-bit for BS_FilSysType FAT12, 16-bit for FAT16.
3. Assume 32-bit for all other cases.

This is problematic, because

1. the possible values of BS_FilSysType are not defined in the spec (and
   indeed vary across different filesystem creation tools).
2. The assumption that any other FS type must be FAT32 is incorrect and
   results in futile attempts to mount non-FAT filesystems using the FAT
   driver.

This commit aligns with the parsing algorithm used by the Linux kernel,
borrowing code from fs/fat to ensure maximum compatibility and
robustness. The borrowed code has been kept to a minimum and as close as
possible to upstream, making it easy to pull in bugfixes from future
versions of the Linux kernel. Both efibootguard and the Linux kernel are
GPL-2.0 licensed, permitting this approach.

For an in-depth understanding of the FAT file system, please refer to
http://elm-chan.org/docs/fat_e.html.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
[Jan: style fix, add commenting sector_size = 0]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>