efi: implement systemd boot loader interface
This patch implements the LoaderDevicePartUUID part of the systemd boot
loader interface to pass data from the loader to the OS / systemd. The
data is passed via EFI variables which are set by the first-stage loader
(the one on the ESP), or alternatively by the first loader that is
executed. By that, userspace components can later inspect this variable
to e.g. limit the search for config partitions to the device it was
bootet from. Currently only the LoaderDevicePartUUID is implemented.

Technically, the loader asks the EFI API for the UUID of the partition it
is executed from. Normally that is the ESP partition. Then, this UUID is
assigned to the LoaderDevicePartUUID EFI variable (in case not set yet).
This logic is crucial to correctly support chain-loading uses-cases and
also aligned with how systemd boot implements this.

For the sake of completeness, this logic is also added to the efi stub.
When using it in combination with the EBG loader, this is irrelevant,
but when starting the UKI directly it is needed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>