Use --no-warn-rwx-segments only when needed
GNU binutils 2.39 introduced a warning for LOAD segments with RWX
permissions. Due to the default Werror, the build fails for non-
x86 machines when using binutils >= 2.39 and gnu-efi <= 3.0.15.
x86 machines didn't have a faulty linker script in the first place.

The gnu-efi commits 98f1250 and 231c763 fix the need for having
--no-warn-rwx-segments on riscv and arm, respectively, and are
included in gnu-efi 3.0.16

Do only apply the linker flag --no-warn-rwx-segments for non-x86
and gnu-efi < 3.0.16 if that flag is supported (binutils >= 2.39).

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>