bg_gen_unified_kernel: Detect and reject signed stub images
Someone decided to sign the kernel stub. The generator does not expect
that and delivers in inconsistent UKI as output. We could address that
by removing the stub signature, but it is simply not worth it, given
that signing the stub never makes sense. Instead, detect the unexpected
input and simply reject it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>