Prioritize volumes on boot media
This ensures that kernels are preferably loaded from the boot medium in
case there are multiple partitions with identical label names. For that
purpose, move all volumes on the boot medium to the front of the list.

As we are potentially reordering the list while building it, we need to
postpone its printing to the end when it is fully constructed.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>