Rollback in TESTING state when kernel image cannot load
In an update scenario, if ustate was just switched from
INSTALLED=1 to TESTING=2 but the kernel image to be tested
is not present, reboot to trigger a fallback into the
original boot path.

This is a simple optimization to not wait until the watchdog
kicks in and reboots the system to the same effect. Incidentally,
this also help to fix downstream issues where U-Boot was incorrectly
configured to feed the watchdog indefinitely, i.e., no reboot happens,
ever.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>