kernel-stub: only warn if device part uuid variable cannot be set
We got reports from the field stating that the boot failed because the
efi variables could not be set and we bail out in this case. As the
systemd boot protocol is not a mandatory feature, we can safely ignore
this case. This is in line with the implementation in systemd-boot.

We now change the bail-out to an error message without bail-out.

Fixes: 7c90e82 ("efi: implement systemd boot loader interface")
Reported-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>