Align EFI linking options with recent gnu-efi
Add linker and objcopy switches that gnu-efi uses since 3.0.19 and that
were missing in our configuration so far. Those have no negative impact
with older gnu-efi and toolchains - tested down to Debian 10 setups -
while fixing some gaps with gnu-efi 3.0.19. These gaps break validation
of signatures with EFI runtimes, specifically U-Boot. Seen on armhf.

There are two exceptions in the alignment:

We do not replace -shared with -pie --no-dynamic-linker as that is
breaking with older build environments while -shared is fine for all
versions.

And then we need to keep .dynsym for 3.0.15 and older because arm64 is
otherwise breaking.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>