Do exploit constructors for registering drivers
Unfortunately, it turned out that we are creating holes in the final EFI
binaries by using sections that are not know to the gnu-efi linker
script. These wholes cause troubles when signing and specifically later
on verifying the signatures.

Avoid this by moving to a classic driver registration pattern where
constructor functions call a central services which adds the passed
driver descriptor to a list. This list is then later on walked for the
actual probing.

That works nicely for gnu-efi 3.0.16 and later after it added support
for calling constructors during init. For older versions, we have to do
that ourselves, though. We are using part of the old pattern for it
which annotated start and end of the init_array and manually walked it,
calling the listed functions.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>