fix: resolve segfault when building with busybox/ash
This patch resolves a segmentation fault caused by the configure script, when building efibootguard using the BusyBox shell ash. The segmentation fault seems to be caused by the BusyBox shell consuming the opening $(( as the beginning of an arithmetic statement.

To mitigate this, this patch simply inserts one space character between the opening double parenthesis, whereby the BusyBox shell no longer tries to parse this as an arithmetic expression.

Signed-off-by: Philip Schildkamp <philip.schildkamp@uni-koeln.de>